requirements.txt
numpy
pandas
matplotlib
seaborn
opencv-python
numba
google-generativeai
python-dotenv